requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-Restless==0.17.0
Flask-Compress==1.2.1
requests==2.7.0
manage.py==0.2.10
FormEncode==1.3.0a1
Jinja2==2.7.2
MarkupSafe==0.21
PyYAML==3.10
Unidecode==0.04.14
Werkzeug==0.9.4
future==0.12.3
gevent==1.0
greenlet==0.4.2
grequests==0.2.0
itsdangerous==0.24
python-dateutil==2.2
python-slugify==0.0.7
six==1.3.0
loremipsum==1.0.2
flask-cors==2.1.0
future==0.15.0
loremipsum==1.0.5
redis==2.10.3
rq==0.4.6
